+++ lib/rubocop/runner.rb
@@ -196,9 +196,9 @@ module RuboCop
 
     def considered_failure?(offense)
       # For :autocorrect level, any offense - corrected or not - is a failure.
-      return true if @options[:fail_level] == :autocorrect
-
       return false if offense.disabled?
+
+      return true if @options[:fail_level] == :autocorrect
 
       !offense.corrected? && offense.severity >= minimum_severity_to_fail
     end